+++ contracts/os/andromeda-kernel/src/state.rs
@@ -43,7 +43,9 @@ pub const IBC_FUND_RECOVERY: Map<&Addr, Vec<Coin>> = Map::new("ibc_fund_recovery
 pub const PENDING_MSG_AND_FUNDS: Item<Ics20PacketInfo> = Item::new("pending_execute_msg");
 
 /// Used to store sequence/channel against an ExecuteMsg, to be sent after an ack of ICS20
-pub const CHANNEL_TO_EXECUTE_MSG: Map<String, Ics20PacketInfo> = Map::new("channel_to_execute_msg");
+/// Key is (channel_id, sequence)
+pub const CHANNEL_TO_EXECUTE_MSG: Map<(String, u64), Ics20PacketInfo> =
+    Map::new("channel_to_execute_msg");
 
 /// Used to temporarily store the most recent ExecuteMsg with the corresponding Coin to be sent in a reply for ICS20 transfer
 pub const REFUND_DATA: Item<RefundData> = Item::new("refund_data");